+++ CHANGELOG.md
@@ -11,6 +11,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 
+## 1.1.0 - 2019-03-26
+### Changed
+
+- GaugeHelper was renamed to GaugeMetric and is now obtained from the
+  MetricsConfiguration singleton. It was simplified to only deal with Integers.
+  The only method to call is 'gaugeHelper.incremementMetric(key)' to increment
+  a value. If a metric key is non-null, it is sent to Graphite and the metric
+  key value is reset to null.
+
+## 1.0.2 - 2019-03-23
+### Added
+
+- GaugeHelper to help with instantaneous data
 
 ## 1.0.1 - 2018-08-13
 ### Changed